Add files via upload
1.Write a program to calculate the total bill tax amount for a list of billing amounts 
passed as an array of long integers.
Up to the amount 1000, there is no tax applicable, subsequently, a flat tax of 10% 
is applicable for the remaining amount as per the tax rate
Note:
All calculations and results should be integer-based ignoring fractions
You are expected to write code int the calcTotalTax function only which will receive 
the first parameter as the number of items in the array and the second parameter 
is the array itself. You are not required to take input from the console.