Sieve of Eratosthenes
The algorithm implemented in this  Java program is the Sieve of Eratosthenes. The Sieve of Eratosthenes is an ancient algorithm for finding all prime numbers up to a given limit n. It works by iteratively marking the multiples of each prime, starting from 2. The unmarked numbers that remain are then considered prime.